+++ app/src/containers/WalletAccounts.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>
+    Hello world Accounts
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'walletAccounts',
+  data() {
+    return {}
+  }
+}
+</script>
