+++ app/src/containers/WalletHistory.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>
+    Hello world History
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'walletHistory',
+  data() {
+    return {}
+  }
+}
+</script>
