+++ app/src/containers/WalletHome.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>
+    Hello world Home
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'walletHome',
+  data() {
+    return {}
+  }
+}
+</script>
