+++ app/src/containers/WalletSettings.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>
+    Hello world Settings
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'walletSettings',
+  data() {
+    return {}
+  }
+}
+</script>
